% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_some_missing.R
\name{check_some_missing}
\alias{check_some_missing}
\title{Check if some missing values are present}
\usage{
check_some_missing(x)
}
\arguments{
\item{x}{the vector to check}
}
\value{
TRUE or FALSE
}
\description{
Check if some missing values are present, but not all are missing.
returns a boolean. This check is done to save time for vectors where filling
is not needed
}
