% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce2.R
\name{coalesce2}
\alias{coalesce2}
\title{coalesce missing values}
\usage{
coalesce2(...)
}
\arguments{
\item{...}{vectors. All imputs should have the same length}
}
\value{
vector with the same length as the first vector
}
\description{
Returns the first non null column element for each row. code from:
https://stackoverflow.com/questions/19253820/how-to-implement-coalesce-efficiently-in-r
}
\examples{
coalesce2(c(NA,1), c(2, NA))
}
