% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_previous.R
\name{fill_missing_previous}
\alias{fill_missing_previous}
\title{Fill missing previous}
\usage{
fill_missing_previous(x, min_known_n = NULL, min_known_p = NULL)
}
\arguments{
\item{x}{The vector to fill}

\item{min_known_n}{numeric value: the minimum number of not-missing values}

\item{min_known_p}{numeric value between 0 and 1: the minimum fraction of not-missing values}
}
\value{
a filled vector
}
\description{
Fill all missing values in a vector with the previous value if it is known.
}
\examples{
fill_missing_previous(c(1, 2, NA))
fill_missing_previous(c(NA, 1, 2, NA))
}
