% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_strict.R
\name{fill_missing_strict}
\alias{fill_missing_strict}
\title{Fill missing strict}
\usage{
fill_missing_strict(x)
}
\arguments{
\item{x}{The vector to fill}
}
\value{
a filled vector
}
\description{
Fill all missing values in a vector with the same value if it is known. Only
fills the value when all known values are the same
}
\examples{
fill_missing_strict(c(NA, 1))
}
