% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_value.R
\name{fill_value}
\alias{fill_value}
\title{fill missing value}
\usage{
fill_value(x, value)
}
\arguments{
\item{x}{vectors. All inputs should have the same length}

\item{value}{a value with the same class as x}
}
\value{
vector with the same length as the first vector
}
\description{
Returns a vector with all missing values filled with another value
}
\examples{
fill_value(c(NA,1), 2)
}
