% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_vector.R
\name{fill_vector.max}
\alias{fill_vector.max}
\title{fill_vector.max}
\usage{
fill_vector.max(x, x_na_omit)
}
\arguments{
\item{x}{the vector to be filled}

\item{x_na_omit}{the x vector without na's}
}
\description{
fill_vector.max
}
