% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_vector.R
\name{fill_vector_min}
\alias{fill_vector_min}
\title{fill_vector_min}
\usage{
fill_vector_min(x, x_na_omit)
}
\arguments{
\item{x}{the vector to be filled}

\item{x_na_omit}{the x vector without NA values}
}
\description{
fill_vector_min
}
