% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_strict.R
\name{fill_missing_strict}
\alias{fill_missing_strict}
\title{Fill missing strict}
\usage{
fill_missing_strict(x, min_known_n = NULL, min_known_p = NULL)
}
\arguments{
\item{x}{The vector to fill}

\item{min_known_n}{numeric value: the minimum number of not-missing values}

\item{min_known_p}{numeric value between 0 and 1: the minimum fraction of not-missing values}
}
\value{
a filled vector
}
\description{
Fill all missing values in a vector with the same value if it is known. Onlx_na_omit
fills the value when all known values are the same
}
\examples{
fill_missing_strict(c(NA, 1))
}
